requirementsRAG.txt
python-dotenv
streamlit
langchain-community
langchain-text-splitters
langchain-groq
unstructured
unstructured[pdf]
transformers
sentence-transformers
nltk